Add support for analyzer image references
